simple clean up, and update all.txt

+++ ruby/all.rb
@@ -1,9 +1,9 @@
 ruby_root = File.dirname(__FILE__)
 input_root = File.join(ruby_root, "..", "input")
 
-Dir["#{ruby_root}/day*.rb"].each do |f|
-  f =~ /day(\d+)/
-  daynum = $1
+(1..25).map(&:to_s).each do |daynum|
+  f = "ruby/day#{daynum}.rb"
+  next unless File.exist?(f)
   puts "------", "day #{daynum}"
   system "ruby", f, in: File.open(File.join(input_root, daynum))
 end
